+++ apps/user/src/app/(dashboard)/note/temporary-note/page.tsx
@@ -7,7 +7,7 @@ export default function Page() {
   const todoid = 2;
   const noteid = 3;
   return (
-    <main className='desktop:pl-[360px] tablet:pl-[84px]  tablet:pr-[24px] mobile:pl-[16px] mobile:pr-[16px] bg-slate-100 min-h-screen'>
+    <main>
       <ToastContainer limit={1} />
       배경이 bg-slate-100인 유저 페이지- 대시보드, 모든 할일 목록 페이지, 목표
       상세 페이지, 노트 모아보기 <br />
@@ -16,17 +16,17 @@ export default function Page() {
       <div className='flex flex-row gap-5'>
         <Link
           className='border border-blue-400'
-          href={`/${goalid}/note/list`}
+          href={`/note/${goalid}/list`}
         >{`goalid:1 의 Note 모아보기`}</Link>
         <Link
           className='border border-blue-400'
-          href={`/${goalid}/${todoid}/note/new`}
+          href={`/note/${goalid}/${todoid}/new`}
         >
           {`golaid:2 & todoid:3의 Note 작성하러가기`}
         </Link>
         <Link
           className='border border-blue-400'
-          href={`/${goalid}/${todoid}/note/${noteid}`}
+          href={`/note/${goalid}/${todoid}/${noteid}`}
         >
           {`golaid:2 & todoid:3의 Note 수정 하러가기`}
         </Link>
