+++ apps/user/src/app/layout.tsx
@@ -1,7 +1,6 @@
-import '../globals.css';
+import './globals.css';
 import '@jeiltodo/ui/styles.css';
 import type { Metadata } from 'next';
-import { SidebarUser } from '../../widgets/user';
 import { ToastContainer } from 'react-toastify';
 
 export const metadata: Metadata = {
@@ -17,13 +16,8 @@ export default function RootLayout({
   return (
     <html lang='en'>
       <body className='font-pretendard-regular'>
-        <div className=''>
-          <SidebarUser />
-          <main className='desktop:pl-[360px] tablet:pl-[84px]  tablet:pr-[24px] mobile:pl-[16px] mobile:pr-[16px] bg-slate-100 min-h-screen'>
-            <ToastContainer limit={1} />
-            {children}
-          </main>
-        </div>
+        <ToastContainer limit={1} />
+        {children}
       </body>
     </html>
   );
