--- apps/user/src/app/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function Page() {
-  return (
-    <main className='p-4 '>
-      
-    </main>
-  );
-}
